#104 Extract all tauri app to its own module
