+++ src-tauri/src/app/mod.rs
@@ -1,0 +1,3 @@
+pub mod handles;
+pub mod menu;
+pub mod store;
